+++ src/App.tsx
@@ -1,16 +1,17 @@
 // Redux
 import { Provider } from "react-redux";
-import { store } from "./providers/redux";
+import { PersistGate } from 'redux-persist/integration/react'
+import { store, persistor } from "./providers/redux";
 
 // Containers
 import { Reddit } from "./containers/Reddit";
 
-function App() {
+export default function App() {
   return (
     <Provider store={store}>
-      <Reddit />
+      <PersistGate loading={null} persistor={persistor}>
+        <Reddit />
+      </PersistGate>
     </Provider>
   );
 }
-
-export default App;