+++ src/containers/Reddit/Content/index.tsx
@@ -1,12 +1,15 @@
 // Types
 import styled from "@emotion/styled";
 import { Menu, PostContent } from "../../../components";
+
+import { PostList } from "./PostList";
+import { LoadingSkeleton } from "./LoadingSkeleton";
+
 import {
   IDismissAllPostsAction,
   IDismissPostAction,
   IReadPostAction,
 } from "../actions";
-import { PostList } from "./PostList";
 
 import { IPost } from "../../../dal/types";
 
@@ -16,13 +19,21 @@ export interface ContentProps {
   readPost: (id: string) => IReadPostAction;
   dismissPost: (id: string) => IDismissPostAction;
   dismissAllPosts: () => IDismissAllPostsAction;
+  isServiceLoading: boolean;
+  serviceError: string | null;
 }
 
 export default function Content({
   openedPost,
   dismissAllPosts,
+  serviceError,
+  isServiceLoading,
   ...props
 }: ContentProps) {
+
+  if (serviceError) return <div>{serviceError}</div>
+  if (isServiceLoading) return <LoadingSkeleton />
+
   return (
     <StyledContent>
       <Menu