+++ src/containers/Reddit/actions.ts
@@ -5,6 +5,7 @@ import { RedditActionTypes } from './actionsTypes';
 import { getPosts } from '../../dal';
 // API types
 import { IPost } from '../../dal/types';
+// App Service actions
 
 // Action Definition
 export type IFetchPostsAction = (dispatch: Dispatch<AnyAction>) => Promise<void | {
@@ -26,16 +27,45 @@ export interface IReadPostAction {
     payload: string
 }
 
+// TODO: Move service actions to separate reducer.
+export interface IFetchingAction {
+    type: RedditActionTypes.READ_POST,
+    payload: string
+}
+
+export interface IFetchErrorAction {
+    type: RedditActionTypes.READ_POST,
+    payload: string
+}
+
 export function fetchPostsAction(): IFetchPostsAction {
     return async (dispatch: Dispatch) => {
-        return getPosts().then(data => {
-            dispatch({
-                type: RedditActionTypes.FETCH_DATA,
-                payload: data,
-            })
-        }).catch(err => {
-            console.log(">>>>>", err);
-        });
+        dispatch({
+            type: RedditActionTypes.SERVICE_ERROR,
+            payload: null,
+        })
+        dispatch({
+            type: RedditActionTypes.SERVICE_FETCHING,
+            payload: true,
+        })
+        return getPosts()
+            .then(data => {
+                dispatch({
+                    type: RedditActionTypes.FETCH_DATA,
+                    payload: data,
+                })
+            }).catch(err => {
+                // TODO: Specify error to avoid generic error response.
+                dispatch({
+                    type: RedditActionTypes.SERVICE_ERROR,
+                    payload: 'Failed to fetch',
+                })
+            }).finally(() => {
+                dispatch({
+                    type: RedditActionTypes.SERVICE_FETCHING,
+                    payload: false,
+                })
+            });
     };
 
 }